requirements.txt
streamlit
wordcloud
matplotlib
numpy
requests
pillow